+++ R/props.R
@@ -6,15 +6,11 @@
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
-#' @seealso \code{\link{check_prop}} and \code{\link{check_vector}}
 #' @export
-#'
-#' @examples
-#' check_props(1, error = FALSE)
-#' check_props(1.1, error = FALSE)
-#' check_props(c(0, 1), error = FALSE)
 check_props <- function(x, x_name = substitute(x),
                        error = TRUE) {
+  .Deprecated()
+  
   x_name <- deparse_x_name(x_name)
 
   check_flag_internal(error)